added start of pathfinding

+++ src/game/pathfinding/Pathfinding.java
@@ -5,7 +5,7 @@ import java.awt.geom.Point2D;
 public class Pathfinding {
 
     public static Point2D getEndPoint() {
-        return new Point2D.Double(100, 100);
+        return new Point2D.Double(200, 100);
     }
 
     public static Point2D getStartPoint() {
